maintain test cases of generate/decode TransactionPayload->ScriptFunction hex  in one file
